+++ .github/CHANGELOG.md
@@ -35,6 +35,9 @@ Vincent Michaud-Rioux
 
 ### Improvements
 
+* Replace deprecated InitArguments by InitializationSettings.
+[(#57)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/57)
+
 * Remove deprecated `set-output` commands from workflow files.
 [(#56)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/56)